refactor: split out create-actual-test-expression
